Remove leading slashes from asset paths
The trailing slash is already included on the definiton of the `ELEMENTOR_PRO_ASSETS_URL` which results in the compiled URL having a double slash. this is fine in environments like Apache where double slashes are normalised, but for environments which don't tolerate that (e.g. S3) it causes the asset to 404.